requirements.txt
altair==4.2.0
anyio==3.6.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.88.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
h11==0.14.0
idna==3.4
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.1.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.1
starlette==0.22.0
streamlit==1.15.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
util2==0.0.0.4
utils==1.0.1
utils2==0.1.2
uvicorn==0.20.0
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
xgboost==1.7.1
zipp==3.11.0
